pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cogito"
version = "0.1.0"
description = "Thinking Tools Framework for AI-Augmented Development"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "Thinking Tools Framework Team"},
]
keywords = [
    "ai", "thinking-tools", "metacognition", "development-tools",
    "claude-code", "mcp", "ai-first"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    "jinja2>=3.1.0,<4.0.0",
    "pyyaml>=6.0.0,<7.0.0",
    "jsonschema>=4.20.0,<5.0.0",
    "watchdog>=3.0.0,<4.0.0",
    "pydantic>=2.5.0,<3.0.0",
    "fastmcp>=2.13.0",
    "click>=8.1.0,<9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]

[project.scripts]
cogito = "cogito.ui.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/thinking-tools-framework"
Documentation = "https://thinking-tools-framework.readthedocs.io"
Repository = "https://github.com/yourusername/thinking-tools-framework"
Issues = "https://github.com/yourusername/thinking-tools-framework/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/cogito"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--cov=cogito",
    "--cov-report=html",
    "--cov-report=term-missing",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
explicit_package_bases = true
namespace_packages = true
mypy_path = "src"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = []

[tool.black]
line-length = 100
target-version = ['py311']
